requirements.txt
# FastAPI Core
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.1.0
schedule==1.2.0
pydantic>=2.11.7,<3.0.0

# MCP Integration
fastmcp==2.12.3

# AI Services
requests>=2.28.0
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0